+++ Source/Model/DetailText.swift
@@ -1,9 +1,9 @@
 //
-//  SectionSpec.swift
+//  DetailText.swift
 //  QuickTableViewController
 //
-//  Created by Ben on 18/01/2016.
-//  Copyright © 2016 bcylin.
+//  Created by bcylin on 31/12/2018.
+//  Copyright © 2018 bcylin.
 //
 //  Permission is hereby granted, free of charge, to any person obtaining a copy
 //  of this software and associated documentation files (the "Software"), to deal
@@ -24,22 +24,37 @@
 //  SOFTWARE.
 //
 
-import Nimble
-import Quick
-import QuickTableViewController
+import UIKit
 
-internal final class SectionSpec: QuickSpec {
+/// An enum that represents a detail text with `UITableViewCell.CellStyle`.
+public enum DetailText: Equatable {
 
-  override func spec() {
-    describe("initialization") {
-      let row = NavigationRow(text: "", detailText: .none)
-      let section = Section(title: "title", rows: [row], footer: "footer")
-      it("should initialize with given parameters") {
-        expect(section.title) == "title"
-        expect(section.rows).to(haveCount(1))
-        expect(section.rows.first as? NavigationRow) == row
-        expect(section.footer) == "footer"
-      }
+  /// Does not show a detail text in `UITableViewCell.CellStyle.default`.
+  case none
+  /// Shows the detail text in `UITableViewCell.CellStyle.subtitle`.
+  case subtitle(String)
+  /// Shows the detail text in `UITableViewCell.CellStyle.value1`.
+  case value1(String)
+  /// Shows the detail text in `UITableViewCell.CellStyle.value2`.
+  case value2(String)
+
+  /// Returns the corresponding table view cell style.
+  public var style: UITableViewCell.CellStyle {
+    switch self {
+    case .none:     return .default
+    case .subtitle: return .subtitle
+    case .value1:   return .value1
+    case .value2:   return .value2
+    }
+  }
+
+  /// Returns the associated text of the case.
+  public var text: String? {
+    switch self {
+    case .none:
+      return nil
+    case let .subtitle(text), let .value1(text), let .value2(text):
+      return text
     }
   }