+++ Source/Protocol/Configurable.swift
@@ -36,8 +36,8 @@ public protocol Configurable {
 extension UITableViewCell {
 
   internal func defaultSetUp(with row: Row & RowStyle) {
-    textLabel?.text = row.title
-    detailTextLabel?.text = row.subtitle?.text
+    textLabel?.text = row.text
+    detailTextLabel?.text = row.detailText?.text
 
     // Reset the accessory view in case the cell is reused.
     accessoryView = nil