+++ Tests/Model/SubtitleSpec.swift
@@ -23,7 +23,6 @@
 //  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 //  SOFTWARE.
 //
-
 import Nimble
 import Quick
 @testable import QuickTableViewController
@@ -31,81 +30,12 @@ import Quick
 internal final class SubtitleSpec: QuickSpec {
 
   override func spec() {
-    describe("subtitle style") {
-      it("should return the descriptive name of the style") {
-        expect(Subtitle.none.style.stringValue) == ".default"
-        expect(Subtitle.belowTitle("text").style.stringValue) == ".subtitle"
-        expect(Subtitle.rightAligned("text").style.stringValue) == ".value1"
-        expect(Subtitle.leftAligned("text").style.stringValue) == ".value2"
-      }
-    }
-
-    describe("associtated value") {
-      let none = Subtitle.none
-      let belowTitle = Subtitle.belowTitle("text")
-      let rightAligned = Subtitle.rightAligned("text")
-      let leftAligned = Subtitle.leftAligned("text")
-
-      it("should return the associated text") {
-        expect(none.text).to(beNil())
-        expect(belowTitle.text) == "text"
-        expect(rightAligned.text) == "text"
-        expect(leftAligned.text) == "text"
-      }
-    }
-
-    describe("equatable") {
-      context(".none") {
-        it("should be equal when both are .none") {
-          let a = Subtitle.none
-          let b = Subtitle.none
-          expect(a) == b
-        }
-      }
-
-      context(".belowTitle") {
-        let a = Subtitle.belowTitle("Same")
-        let b = Subtitle.belowTitle("Same")
-        let c = Subtitle.belowTitle("Different")
-        let d = Subtitle.rightAligned("Same")
-        let e = Subtitle.none
-
-        it("should be equal only when type and associated value match") {
-          expect(a) == b
-          expect(a) != c
-          expect(a) != d
-          expect(a) != e
-        }
-      }
-
-      context(".rightAligned") {
-        let a = Subtitle.rightAligned("Same")
-        let b = Subtitle.rightAligned("Same")
-        let c = Subtitle.rightAligned("Different")
-        let d = Subtitle.leftAligned("Same")
-        let e = Subtitle.none
-
-        it("should be equal only when type and associated value match") {
-          expect(a) == b
-          expect(a) != c
-          expect(a) != d
-          expect(a) != e
-        }
-      }
-
-      context(".leftAligned") {
-        let a = Subtitle.leftAligned("Same")
-        let b = Subtitle.leftAligned("Same")
-        let c = Subtitle.leftAligned("Different")
-        let d = Subtitle.belowTitle("Same")
-        let e = Subtitle.none
-
-        it("should be equal only when type and associated value match") {
-          expect(a) == b
-          expect(a) != c
-          expect(a) != d
-          expect(a) != e
-        }
+    describe("compatibility") {
+      it("should return the corresponding detail text") {
+        expect(Subtitle.none.detailText) == DetailText.none
+        expect(Subtitle.belowTitle("text").detailText) == .subtitle("text")
+        expect(Subtitle.rightAligned("text").detailText) == .value1("text")
+        expect(Subtitle.leftAligned("text").detailText) == .value2("text")
       }
     }
   }