+++ Tests/Row/NavigationRowSpec.swift
@@ -32,15 +32,15 @@ internal final class NavigationRowSpec: QuickSpec {
 
   override func spec() {
     describe("initialization") {
-      let subtitle = Subtitle.belowTitle("subtitle")
+      let detailText = DetailText.subtitle("subtitle")
       let icon = Icon.named("icon")
 
       var invoked = false
-      let row = NavigationRow(title: "title", subtitle: subtitle, icon: icon, action: { _ in invoked = true })
+      let row = NavigationRow(text: "title", detailText: detailText, icon: icon, action: { _ in invoked = true })
 
       it("should initialize with given parameters") {
-        expect(row.title) == "title"
-        expect(row.subtitle) == subtitle
+        expect(row.text) == "title"
+        expect(row.detailText) == detailText
         expect(row.icon) == icon
         expect(row.cellReuseIdentifier) == "UITableViewCell.subtitle"
 
@@ -54,10 +54,10 @@ internal final class NavigationRowSpec: QuickSpec {
     }
 
     describe("cellReuseIdentifier") {
-      let a = NavigationRow(title: "", subtitle: .none)
-      let b = NavigationRow(title: "", subtitle: .belowTitle(""))
-      let c = NavigationRow(title: "", subtitle: .rightAligned(""))
-      let d = NavigationRow(title: "", subtitle: .leftAligned(""))
+      let a = NavigationRow(text: "", detailText: .none)
+      let b = NavigationRow(text: "", detailText: .subtitle(""))
+      let c = NavigationRow(text: "", detailText: .value1(""))
+      let d = NavigationRow(text: "", detailText: .value2(""))
 
       it("should return the backward compatible strings") {
         expect(a.cellReuseIdentifier) == "UITableViewCell.default"
@@ -69,40 +69,40 @@ internal final class NavigationRowSpec: QuickSpec {
 
     describe("equatable") {
       let image = UIImage()
-      let a = NavigationRow(title: "Same", subtitle: .belowTitle("Same"), icon: .image(image), action: nil)
+      let row = NavigationRow(text: "Same", detailText: .subtitle("Same"), icon: .image(image), action: nil)
 
       context("identical parameters") {
-        let b = NavigationRow(title: "Same", subtitle: .belowTitle("Same"), icon: .image(image), action: nil)
+        let this = NavigationRow(text: "Same", detailText: .subtitle("Same"), icon: .image(image), action: nil)
         it("should be equal") {
-          expect(a) == b
+          expect(this) == row
         }
       }
 
-      context("different titles") {
-        let c = NavigationRow(title: "Different", subtitle: .belowTitle("Same"), icon: .image(image), action: nil)
+      context("different texts") {
+        let this = NavigationRow(text: "Different", detailText: .subtitle("Same"), icon: .image(image), action: nil)
         it("should not be equal") {
-          expect(a) != c
+          expect(this) != row
         }
       }
 
-      context("different subtitles") {
-        let d = NavigationRow(title: "Same", subtitle: .belowTitle("Different"), icon: .image(image), action: nil)
+      context("different detail texts") {
+        let this = NavigationRow(text: "Same", detailText: .subtitle("Different"), icon: .image(image), action: nil)
         it("should not be equal") {
-          expect(a) != d
+          expect(this) != row
         }
       }
 
       context("different icons") {
-        let e = NavigationRow(title: "Same", subtitle: .belowTitle("Same"), icon: .image(UIImage()), action: nil)
+        let this = NavigationRow(text: "Same", detailText: .subtitle("Same"), icon: .image(UIImage()), action: nil)
         it("should not be equal") {
-          expect(a) != e
+          expect(this) != row
         }
       }
 
       context("different actions") {
-        let f = NavigationRow(title: "Same", subtitle: .belowTitle("Same"), icon: .image(image), action: { _ in })
+        let this = NavigationRow(text: "Same", detailText: .subtitle("Same"), icon: .image(image), action: { _ in })
         it("should be equal regardless of the actions attached") {
-          expect(a) == f
+          expect(this) == row
         }
       }
     }